Creation of Texture
Added texture class that takes input of an image and saves it as it's own readable file. This file stores its member variables to allow for optimised loading on next load. The texture class stores colour data in its own format so that override colour can be applied.

Program object is now being tracked in renderer.

Added stb_image for image loading.

Entity adding and duplicating now return the new entity reference, this is so interacting with the new entity is easy.

Created namespace to store file tags as data types, previously stored them as defines however this didnt allow for optimisation.

Added more to Material, it now tracks albedo, metalic, normal and height maps as textures. It also has options for emitting light, specularHighlights and reflections.